added arrowscaler script
