+++ Assets/Scripts/ArrowScalerScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b79a93115b8508d4ca5f404488716120
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
