+++ Assets/Scripts/PauseHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2fcf2a5ebbf810e40aecb644caf900a4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
